move policy view into its own file
